merge: refactor (#46)
* refactor: creation de package

* add: package-info

* fix(name): nom du package

* fix: typo

* feat(diagramme): code source des classes

* feat(diagramme): les diagrammes en png

* fix: retrai d'élément inutile

* fix: compilation plantuml

* docs: deplacement des diagrame de classe

* fix(legend): reduction

* fix: passage au diagrame en svg

* add: ajout des diagrame dans la doc

* fix(doc): lien doc + gitignore

ajout de l'ignore des png est svg de plantUML
ajout du lien de la doc dans le readme

* fix(legend): reduction

L'image ne s'affiche pas.

* fix(name) classes/classes_* => classe/*

* feat(legend): parametrable

* refactor: deplacement des diagrame

* feat: supprestion des diagrame générer

* fix: gitignore des svg

* fix: utilisation des quote et double quote

* fix:  majuscule

* feat(doc): ajout des svg a la javadoc

* docs: mise a jour des diagrame

* refactor: cell mise dans un package

* fix: n'aurais pas due ètre push

---------
