+++ doc/plantuml/classes/Coord.puml
@@ -1,15 +1,16 @@
-@startuml classes_coord
+@startuml Coord
 
-title diagramme de la classe gérant les coordonnées
+title "Diagramme de la classe gérant les coordonnées"
 
-!includesub legend.puml!legend
+!$method=1
+!includesub ../legend.puml!legend
 
-!startsub coord
+!startsub Coord
 class "Coord" AS Coord {
 	{field} - numCell : int
 	==
 	.. Constructeurs ..
-	{method} {static} + new Coord() : Coord
+  {method} {static} + new Coord(void) : Coord
 	{method} {static} + new Coord(String) : Coord
 	{method} {static} + new Coord(int,int) : Coord
 	{method} {static} + new Coord(int) : Coord
@@ -40,14 +41,14 @@ class "Coord" AS Coord {
 	{method} + exist(void) : boolean
 }
 note left of Coord::getId
-	id : Coordonnée d'une case,
-	pour le controleur de l'interface graphique.
-	end note
-	note left of Coord::getLine
-	line & column : Coordonnée d'une case,
-	pour le moteur de jeu (la classe gérant la matrice de jeu).
+	id : "Coordonnée d'une case,
+	pour le controleur de l'interface graphique. "
+end note
+note left of Coord::getLine
+	line & column : "Coordonnée d'une case,
+	pour le moteur de jeu (la classe gérant la matrice de jeu)."
 end note
 !endsub
 
-center footer Sert d'interface entre plusieurs modules n'utilisant pas le même système de coordonnées.
+center footer "Sert d'interface entre plusieurs modules n'utilisant pas le même système de coordonnées."
 @enduml
