+++ doc/plantuml/package/grid.puml
@@ -1,8 +1,10 @@
-@startuml classes_grid
+@startuml grid
 
-title diagramme de classes de la grille du jeu
+title "Diagramme de classes de la grille du jeu"
 
-!includesub legend.puml!legend
+!$class=1
+!$method=1
+!includesub ../legend.puml!legend
 
 !startsub grid
 package "Grid" #lightgrey {
@@ -25,8 +27,8 @@ package "Grid" #lightgrey {
 		{method} + getCoord(void) : Coord
 	}
 	note right of GridIt::getCoord
-		Coordonnées de la
-		dernière Cell envoyé.
+		"Coordonnées de la
+		dernière Cell envoyé."
 	end note
 	java.util.Iterator ().. GridIt
 
@@ -51,26 +53,26 @@ package "Grid" #lightgrey {
 		{method} + serialize() : String
 	}
 	note left of Grid::getCell
-		Cherche une cellule matchant le predicat,
+		"Cherche une cellule matchant le predicat,
 		en se déplacant suivant la fonction (une
-		méthode de Grid.GridIterator).
+		méthode de Grid.GridIterator)."
 	end note
 	note left of Grid::iterator
-		Crée un itérateur sur la grille allant
+		"Crée un itérateur sur la grille allant
 		du début(0,0) à la fin(Kakuro.LAST_COORD()).
 		Le mode block est désactivé
-		et le parcours est dans le sens par défault.
+		et le parcours est dans le sens par défault."
 	end note
 	note left of Grid::serialize
-		Renvoit la liste des cellule, séparer par des '|'.
+		'Renvoit la liste des cellule, séparer par des '|'.
 		Les lignes sont séparer par des '\n'.
 		> Exemple : "|B|B|W(5)|\\n|R(7,n)|W(n)|R(4,1)"
-		> Devient : "||5|\\n7,n|3|4,1|\\n"
+		> Devient : "||5|\\n7,n|3|4,1|\\n"'
 	end note
 	java.lang.Iterable ()-- Grid
 	Grid --* GridIt
 }
 !endsub
 
-center footer La classes stockant les cellules du jeu et les parcourant
+center footer "La classes stockant les cellules du jeu et les parcourant"
 @enduml
