+++ src/main/java/fr/mcgcorp/game/error/package-info.java
@@ -1,6 +1,6 @@
 /**
  * package contenant les classes pour gérer les erreurs du jeu.
  * <br>
- * <img src="doc-files/classes_error.svg" alt="Diagramme de classes du package fr.mcgcorp" width="100%" />
+ * <img src="doc-files/error.svg" alt="Diagramme de classes du package fr.mcgcorp.error" width="100%" />
  */
 package fr.mcgcorp.game.error;
