+++ src/main/java/fr/mcgcorp/game/grid/cell/BlackCell.java
@@ -1,11 +1,11 @@
-package fr.mcgcorp.game.grid;
+package fr.mcgcorp.game.grid.cell;
 
 /**
  * Cellule de base, qui peut être soit une case vide(noir), soit une case de résultat, soit une case a remplir.
  *
  * @author HOUGET Julien, PUREN Mewen, PECHON Erwan
  */
-class BlackCell implements Cell {
+public class BlackCell implements Cell {
   /**
    * Méthode qui retourne la cellule, formater pour la transmission.
    *
