+++ src/main/java/fr/mcgcorp/game/grid/cell/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package gérant les case.
+ * <br>
+ * <img src="doc-files/cell.svg" alt="Diagramme de classes du package fr.mcgcorp.grid" width="100%" />
+ */
+package fr.mcgcorp.game.grid.cell;
