+++ src/main/java/fr/mcgcorp/game/grid/package-info.java
@@ -1,8 +1,6 @@
 /**
  * Package contenant les classes pour gérer la grille de jeu.
  * <br>
- * <img src="doc-files/classes_grid.svg" alt="Diagramme de classes du package fr.mcgcorp" width="100%" />
- * <br>
- * <img src="doc-files/classes_cell.svg" alt="Diagramme de classes du package fr.mcgcorp" width="100%" />
+ * <img src="doc-files/grid.svg" alt="Diagramme de classes du package fr.mcgcorp.grid" width="100%" />
  */
 package fr.mcgcorp.game.grid;
