+++ src/main/java/fr/mcgcorp/game/move/package-info.java
@@ -1,6 +1,6 @@
 /**
  * Package gérant les déplacements des entités.
  * <br>
- * <img src="doc-files/classes_move.svg" alt="Diagramme de classes du package fr.mcgcorp" width="100%" />
+ * <img src="doc-files/move.svg" alt="Diagramme de classes du package fr.mcgcorp.move" width="100%" />
  */
 package fr.mcgcorp.game.move;
