+++ src/main/java/fr/mcgcorp/game/package-info.java
@@ -1,8 +1,9 @@
 /**
  * package général du jeu.
  * <br>
- * <img src="doc-files/classes_game.svg" alt="Diagramme de classes du package fr.mcgcorp" width="100%" />
+ * il contient toute la logique de gestion de la partie
  * <br>
- * <img src="doc-files/classes_coord.svg" alt="Diagramme de classes du package fr.mcgcorp" width="100%" />
+ * <img src="doc-files/game.svg" alt="Diagramme de classes du package fr.mcgcorp.game" width="100%" />
+ * 
  */
 package fr.mcgcorp.game;
